+++ mhep/mhep/dev/tests/views/test_organisation_assessments.py
@@ -6,7 +6,10 @@ from rest_framework import status
 from mhep.users.tests.factories import UserFactory
 
 from ... import VERSION
+from ...models import Assessment
 from ..factories import AssessmentFactory, OrganisationFactory
+
+from .mixins import CreateAssessmentTestsMixin
 
 
 class TestListAssessmentsForOrganisation(APITestCase):
@@ -105,3 +108,62 @@ class TestListAssessmentsForOrganisation(APITestCase):
 
         assert response.status_code == status.HTTP_200_OK
         assert expectedAssessmentCount == len(response.data)
+
+
+class TestCreateAssessmentForOrganisation(CreateAssessmentTestsMixin, APITestCase):
+    """
+    note that more tests are provided by the CreateAssessmentTestsMixin, since they are
+    common with the tests for TestCreateAssessment (for an individual)
+    """
+    def post_to_create_endpoint(self, assessment):
+        organisation = OrganisationFactory.create()
+        organisation.members.add(self.user)
+
+        return self.client.post(
+            f"/{VERSION}/api/organisations/{organisation.pk}/assessments/",
+            assessment,
+            format="json"
+        )
+
+    def test_sets_organisation(self):
+        organisation = OrganisationFactory.create()
+        organisation.members.add(self.user)
+
+        self.client.force_authenticate(self.user)
+
+        new_assessment = {
+            "name": "test assessment 1",
+            "description": "test description 1",
+            "openbem_version": "10.1.1",
+        }
+
+        response = self.client.post(
+            f"/{VERSION}/api/organisations/{organisation.pk}/assessments/",
+            new_assessment,
+            format="json"
+        )
+
+        assert status.HTTP_201_CREATED == response.status_code
+
+        assessment = Assessment.objects.get(pk=response.data["id"])
+        assert organisation == assessment.organisation
+
+    def test_fails_if_not_a_member_of_organisation(self):
+        organisation = OrganisationFactory.create()
+
+        self.client.force_authenticate(self.user)
+
+        new_assessment = {
+            "name": "test assessment 1",
+            "description": "test description 1",
+            "openbem_version": "10.1.1",
+        }
+
+        response = self.client.post(
+            f"/{VERSION}/api/organisations/{organisation.pk}/assessments/",
+            new_assessment,
+            format="json"
+        )
+
+        assert status.HTTP_403_FORBIDDEN == response.status_code
+        assert {"detail": "You are not a member of the Organisation."} == response.json()
